Handle Disconnect, Ping, and Status packets (#410)
Summary
========

The networking layer was not handling Disconnect, Ping, and Status
packages. The first two are DEVp2p and the Status packet is the first
packet that must be sent for the Ethereum Wire Protocol (after the
encrypted and DEVp2p handshakes).

Details
========

Adding Subscribers to Connection struct
-----------------------------------------

First we add `subscribers` to `ExWire.P2P.Connection` struct. For a
while I was resisting doing this since I wanted `P2P` to be a pure
module without having to know about the socket connection. Ideally it
could just respond something like `{:send, packet}` and the `P2P.Server`
would handle the actual communication to the socket.

The current implementation of `P2P`, however, is not pure, and it has
knowledge of the socket (it's part of the `Connection` struct). So we
are including `subscriptions` in there for now, especially since there
is a function (`notify_subscribers`) which assumes the struct will have
the subscribers in there.

Adds a rough implementation of handling `Packet.Disconnect,
`Packet.Ping`, and `Packet.Status`:
-----------------------------------

* `Disconnect` is a `DEVp2p` message. It can be sent by a peer at any
point after the encrypted handshake. It signifies that a disconnect is
imminent and that we should disconnect. So we should terminate the TCP
connection.

* `Ping` (and `Pong`) are `DEVp2p` messages that should not be delegated
to the subscribers. That is, they are related to the networking side of
things (ie. "is this node alive?") and not to the blockchain side of
things (e.g. "GetBlockHeaders").

* `Packet.Status` is an eth wire protocol packet, and as such, it is
possible it should not reside in `P2P` but that it should live with the
rest of the packets that are delegated to the subscribers. But since it
is the first packet that must come from all eth wire packets, and since
it requires a response, we include it here for now. It's also important
to note that the status message requires fields like `total_difficulty`
and `best_hash`, values which we do not yet know about our blockchain,
so we are simply mirroring what we get from the other node for now.

Other changes
-------------

* Makes PeerSupervisor (previously a regular supervisor) into a
DynamicSupervisor. This means it can dynamically start children nodes,
which will be outbound `ExWire.P2P.Server`s. We add a
`PeerSupervisor.start_child/1` function that takes in a peer enode url
to start the outbound server.