Turns EVM.address => Account.Address in interface
Block number 177610 in the mainnet has an account that is pushed into
the stack as an integer. The account when represented as an integer does
not have a leading zero, but when represented as hex, it has `00` as its
first two characters.

This led to a subtle bug, where the balance of the account in question
seemed to be zero, when in reality we were checking the wrong account's
balance.

To fix this issue, we ensure that the addresses coming from the EVM (via
the account interface) are all turned into
`Blockchain.Account.Addresses` which must have `160` bits. If they are
short, as was the case in the bug, we pad them.

Additionally:
------------

The SubState's touched accounts should have the binary format for the
address of an account. The `create` system operation was passing the
result of turning an address and nonce into an unsigned integer into the
substate. It should have been passing the `exec_env.address` instead.

Co-authored-by: Matthew Sumner <matt.m.sumner@gmail.com>