Implement EIP-658 (#405)
* Implement EIP-658

Summary
========

Adds a transaction status of `0` for failure and `1` for success. This
transaction status is included in the transaction receipts if the block
number is past the Byzantium hardfork.

More detail
===========

* We update `Transaction.execute`, as well as `MessageCall.execute` and
`ContractCreate.execute` to return the status specified as gamma^z in
section 6, Transaction Execution, of the Yellow Paper (0 for failure, 1
for success).

* After the Bynzatium fork, the transaction receipt's state root should
include a `0` if the transaction failed or a `1` if it succeeded. This
is EIP 658.