Create common_tests.run mix task
Creates `mix common_tests.run` mix task to run a single blockchain test
file for a fork.

The advantages of creating this helpful script as a `mix` task (rather
than just creating it as we have created other scripts) is that it
allows us to have some clear documentation.

- `mix help` will list all the task available in the project, including
`common_tests.run`. It will use the `@shortdoc` for that.

- `mix help common_tests.run` will give a full description of the mix
task, using the information found in `moduledoc`.