Create common_tests.run mix task (#444)
What?
=====

Creates `mix common_tests.run` mix task to run a single blockchain test
file for a fork.

This task uses much of the same logic as does the blockchain_test.exs test. 
So we extract a BlockchainTestRunner module to house all that common 
logic related to running a single blockchain test file.

Why?
=====

When we debug failing blockchain tests, it is often hard to run a single test.
It is possible to do so by modifying certain pieces of the 
`test/blockchain_test.exs` file, but that requires some knowledge of the 
structure of the test setup, and it is both cumbersome and time consuming. 
With this new mix task, we can simply specify the file we need to run and 
it will run that test.

A note on mix task
============

The advantages of creating this helpful script as a `mix` task (rather
than just creating it as we have created other scripts) is that it
allows us to have some clear documentation.

- `mix help` will list all the task available in the project, including
`common_tests.run`. It will use the `@shortdoc` for that.

- `mix help common_tests.run` will give a full description of the mix
task, using the information found in `moduledoc`.