Update blockchain_test pattern matching (#445)
The `BlockchainTestRunner` now returns test results with `{:pass,
result}` or `{:fail, result}`, where `result` is equivalent to what used
to be returned from the same logic before `BlockchainTestRunner` was
extracted. The filtering logic for test failures was updated correctly
to expect the `{:fail, _}` tuple, but the error message constructor was
not updated to expect it. We update that here.