Save state every 1000 blocks in SyncWithInfura
Before we were only saving the state when validation failed. This was
fine early on when failures were more common but now validation errors
are few and father between. Now we save state periodically so if the
task fails due to a catastrophic failure (such as a memory overflow) we
won't have to resync.