Implement EIP-1234
Summary
=======

This implements EIP-1234 which updates the difficulty calculation and
the block reward for Constantinople.

More details
===========

Thus far we have been using Parity's chain config files to generate
configuration for a chain. We update those files to include the values
for Constantinople.

In this change, however, Parity updates not only some values in the json
configuration but also how they are stored. In particular, the block
rewards and the bomb delay factor can now be a single value or json
objects whose key is the block_number of the fork in question and
the value is the thing looked for. So, for example, we could find this
in a block_reward,

```
block_reward: {
 "0" => 300000000,
 "1700000" => 2000000
}
```

We make updates to our `Chain` module to account for this.

Update ethereum_common_tests

Run generate_state_tests

We update the state tests with tests that are still failing and the
passing percentages in the main README.

This change also updates chains from parity:

https://github.com/paritytech/parity-ethereum/tree/master/ethcore/res/ethereum