Turn EVM.Configuration into a behaviour (#451)
What?
=====

We turn `EVM.Configuration` into a Behaviour instead of a protocol.

Why?
=====

Behaviours and Protocols are two ways to implement polymorphism in
Elixir. When we want a module to behave the same way as other modules,
we want it to satisfy a behaviour (e.g. `GenServer`). When we want to
create a somewhat generic set of operations on different types, we use
protocols (e.g. `Enumerable` working on lists, ranges, etc).

Since the `EVM.Configuration` lands on the first camp, we should use a
`Behaviour` for it. This has the side-effect of cleaning up some
dialyzer errors that made no sense previously, such as an `Atom` not
having the `EVM.Configuration` protocol implemented.

Co-authored-by: Matthew Sumner <matt.m.sumner@gmail.com>