Simulate coinbase reward/clean-up in state tests
What changed?
============

Introduces a step in our state tests to create the
miner/beneficiary/coinbase account immediately before setting up
accounts in the `pre` state. It also adds a step after the transaction
execution to cleanup the miner account if we are past EIP-161 (since
that is when we start cleaning up empty accounts).

Why is this needed?
===================

It turns out the state tests are not supposed to be run completely in a
vacuum. Other implementations' test runners duplicate some of their
blockchain implementation's functionality in order to make the state
tests run more like blockchain tests.

There seems to be confusion as to why this is in the first place, but a
[this gitter
conversation](https://gitter.im/ethereum/tests?at=5b85004aff445156164b685f)
is a helpful look at why it happened.

We have state tests that are failing because of state root hash
mismatches: these tests should be simulating giving a miner reward, and
if the environment gives a block number that is past EIP-161, the test
runner should clean up that account (if empty) before comparing state
root hashes. Currently, we have some of these tests passing
coincidentally because we never create a coinbase account, so we
generate a state root hash that matches that of those tests that expect
the coinbase to be cleaned up.

Co-authored-by: Matthew Sumner <matt.m.sumner@gmail.com>