Make BlockInterface into a behaviour
Description
===========

Makes the `EVM.Interface.BlockInterface` into a behaviour, and updates
`Blockchain.Interface.BlockInterface` and
`EVM.Interface.Mock.MockBlockInterface` into modules to satisfy that
behaviour.

While we're doing this, we take the opportunity to rename the modules
since `Interface` does not seem to fit as well. We rename
`BlockInterface` as `BlockHeaderInfo` since the behaviour is about
retrieving information from the block header.

We also create some helper functions in `EVM.BlockHeaderInfo` to make
dynamic dispatch easier.

Why?
---

We were using a protocol to abstract the notion of what a
`BlockInterface` is. But a protocol is not the correct abstraction for a
polymorphic module that responds to the same functions. For that we want
to use behaviours.

Other notes
-----------

While doing this update, we noticed that the function
`get_block_by_hash` in the protocol was not used at all (only privately
in one of the implementations). So we remove that from the protocol (now
behaviour).