Turn AccountInterface into AccountRepo (behaviour)
Description
============

`AccountInterface` was a protocol. We change it to a behaviour, since we
want a module that responds to the same set of functions and Elixir
behaviours are the right tool for that job.

While changing it to a behaviour, we rename it to get rid of the notion
of interface. The module is primarily interested in interacting with
accounts, so we call it an `AccountRepo`.

The two implementations were changed:

- `EVM.Interface.Mock.MockAccountInterface` -> `EVM.Mock.MockAccountRepo`
- `Blockchain.Interface.AccountInterface` -> `Blockchain.Account.Repo`

The `Blockchain.Interface.AccountInterface` had multiple functions that
had the same name inside the module and again inside the protocol
implementation for the module, leading to some strange dependencies
where the protocol function would use the regular module function. To
correct this, those functions were combined into a single function that
satisfies the AccountRepo behavior outlined by the callback.