Fix dialyzer warnings for AccountRepo implementations
Description
===========

We start by removing the ignored warnings for `blockchain/account/repo`
and `evm/mock/mock_account_repo`. This triggered a cascade of fixes that
spanned more modules that just those two. In fact, some warnings led us
to understand that some of the functions were not used at all:

Removals
--------

- Remove add_wei from EVM.AccountRepo. It is not called from the evm
app and thus does not need to be part of the behaviour. It remains as
part of the `Blockchain.Account.Repo`.

- Remove AccountRepo.new_contract_address callback. It is not used.

- Remove unused AccountRepo.message_call callback. It is not used.

Other changes
-------------

- Update correct typespec from EVM.state() to Trie.t() in several places
in `Account.Repo`.

- Correct the return type of EVM.AccountRepo.transfer

- We expose better types from `Account.Repo.Cache` to be used by
`Account.Repo`.

- We address several incorrect types in `Account.Repo.Cache` that
triggered several dialyzer warnings in `Account.Repo`: the Cache's types
are fairly inaccurate in their representation of all the `nil`
possibilities. Based on how the cache is used from the `Account.Repo`
module, we make some updates that more accurately represent reality. For
example, a function in `Account.Repo` calls the caching to update
updating an account with `{nil, nil}` in order to "delete" the account
from cache.

This behavior seems oddly intentional based on a failing test suggesting
that we should not retrieve the account which has been deleted from
storage since that would ignore the fact that it was deleted (but not
yet persisted).

We should find a better way to handle that rather than having to do a
manual insert of `{nil, nil}` to signify "this account has been deleted
in cache, so do not look for it in storage".