Add Blockchain.Account.Inspector (#508)
Adds a `Blockchain.Account.Inspector` module to more easily inspect the
contents of an account. This is mostly for debugging purposes, and it is
not expected that it would be used in production. It takes an account
address (hex) and prints out the balance, nonce, storage root, and code
hash.