Rename common_tests.run task to blockchain_tests.run
The mix task for running a single blockchain test was created first, and
since the task lives in the blockchain app, it was named
`common_tests.run`. But now that we also have a `state_tests.run` mix
task, the `common_tests.run` task name is confusing. We update it to be
`blockchain_tests.run`.