Derive current block from tree in sync_with_infura (#516)
We have been saving both a "current_block_hash" and "current_block_tree"
in the database every 1000 blocks. The problem comes when a failure
occurs at a block that is a multiple of 1000. In that case, it appears
as though we have a corrupted db, but in reality all that has happened
is that we saved a "current_block_hash" too early (before actually
validating the block), and next time we search for that block in our
database, it cannot be found.

So we do two things here to improve that,

1. We derive the current block hash from the current block tree. Since
the tree has the best block in it, that is all we need.

2. We move the saving of the current block tree to after we have
verified and added the block (instead of before). That way, when we try
to run the script again, we will have the correct version of the tree
saved.