Implements the DAO hardfork
Description
===========

Adds a `Blockchain.Hardfork.Dao.execute` to handle the logic of the DAO
hardfork. It transfers the balance of all the DAO accounts into the
beneficiary account (the refund contract).

We follow the pattern used in Geth and Parity. This hardfork in
particular seems a little hacky since it does not behave similar to
other hardforks where the behavior changes. In this hardfork we simply
transfer balances from several accounts to the refund account.

We perform the dao hadrfork before the block adds transactions. This
only gets executed if the chain configuration has dao hardfork data,
meaning it supports the dao hardfork.

We also add header validation for blocks within the [DAO_Fork ...
DAO_Fork + 9] range. Those headers need to have the extra data
"dao-hard-fork" in them in order to be valid.

Tests
------

We add HomesteadToDaoAt5 transition tests (common tests). In order to
test that we have to add a new chain "dao_hardfork_test.json" that
expects the dao transition to occur at block 5. It also includes the
accounts required for the DAO fork and the rest of the configuration is
just like the Homestead test chain.