Improve EVM.Logger output (#531)
A lot of our debugging is happening as we compare stack traces against
geth's vm traces found in Etherescan.

To that end, we make the following changes so that debugging is easier:

1. Add a `step` key to `MachineState` so that we know how many cycles
the vm has executed. This is very helpful in comparing vm traces
line-by-line against geth, since they keep a step count. We increment
the step at the end of each `VM.cycle`.

2. Log the program counter as a number. For some reason, we were logging
the program counter as hex. Etherscan shows it as a number, and it is
much easier to reason about.

3. Log stack as hex. Etherescan shows the stack as hex numbers, not as
integers. Comparing the stack against that will be much easier and will
no longer require manual translation of values.

4. We update the EVM readme to suggest placing the EVM.Logger calls
closer to the beginning of the `cycle` function so that values more
closely match Etherscan's values (e.g. gas used would be equivalent to
next cycle's gas in Etherscan if it's printed at the end of the cycle).