Simplify extcodecopy conditional logic (#535)
We were failing to sync past block 2325632 on the mainnet because
transaction
`0xec32243bad69039271ae01362ec4916be9fbf546e5356db420b1267ac1a038a3` has
an `extcodecopy` operation with positive `size` but the code is empty.

In order to fix that, we remove the conditionals on `extcodecopy` that
returned an unchanged `machine_state`. That conditional tested three
possibilities with logical ORs, meaning any of these three would return
an unchanged `machine_state`:

1. size == 0 -> this is already handled by the Memory.write operation,
so we remove the condition.

2. size + mem_offset > EVM.max_int() -> reading the yellow paper, it
seems that this condition was wrong. Under `excodecopy` in the
appendices outlining Environmental Information, it says:

> The additions in μs[2] + i are not subject to the 2^256 modulo.

3. (code_offset == 0 && account_code == "") -> this condition was the
one failing the transaction mentioned above because even though the code
offset was `0` and the account code was empty, the size was still
positive, meaning we actually needed to write into memory. We add a test
for this case to prevent future regression of this condition.