Make EVM.Configuration simpler (#536)
The problem
============

The EVM.Configuration files are full of repetitive implementations of
functions that simply call the underlying key in a struct,

```elixir
def contract_creation_cost(config), do: config.contract_creation_cost
```

or functions that use a fallback config to call another module's
implementation,

```elixir
def extcodesize_cost(config), do: Frontier.extcodesize_cost(config.fallback_config)
```

This is a lot of duplication both in the functions we write but also in
the calling code. We are doing the following a lot,

```elixir
Evm.Configuration.for(config).contrac_creation_cost(config)
```

when we could simply be doing,

```elixir
config.contract_creation_cost
```

In addition, it is currently difficult to understand the full
configuration for a hardfork as a single snapshot, since a configuration
like Byzantium can have SpuriousDragon's configuration nested within it,
and it have TangerineWhistle nested within it, and it have Homestead,
and it have Frontier:

```elixir
%Configuration.Byzantium{
 # fields
 fallback_config: %SpuriousDragon{
   # fields
   fallback_config: %TangerineWhistle{
     # fields
     fallback_config: %Homestead{
       # fields
       fallback_config: %Frontier{
         # fields
       }
     }
   }
  }
}
```

The solution
============

Since most of the configuration options are just static values, like
boolean flags or some integer, we can simply use a "callback config" at
the moment we are generating the structs. So for example, the `Homestead`
`defstruct` definition can call `Frontier.new()` and override the values
needed to generate its own struct.

This has the benefit of removing repetition in function definitions
(with two exceptions), it improves ease of legibility of a configuration
since we only have a single struct instead of having structs within
structs within structs, and it also opens the possibility of future
configurations that a user may want to define which do not match any
single hardfork.