Add a CLI for Syncing
This patch begins to add a CLI for syncing the blockchain. We create a new sub-app `CLI` to host command-line functions, which is currently limited to syncing. Over time, we will expand this CLI to sync, as well as host an RPC server, etc. We allow syncing from HTTP or IPC (e.g. Infura or a local node).