Better Progress Bar
This patch adds the CLI patch with a better progress bar (notably, it shows the number of blocks synced, not just a percentage). This currently depends on a third-party GitHub fork, so either we'll wait for that fork to resolve to master or clone it ourselves. In either case, this is a placeholder.