Mekle Patricia Trie Cache (#542)
* implement CachingStorage

We need at least two levels of cache. Right now we have one level.

Now we're committing changes after every transaction. This is ok if there are no invalid blocks. But in real network we will be receiving invalid blocks and we shouldn't write their transaction data to db.

The idea is to have CachingTrie: all storage modification will be made to in-memory trie and saved to disk after successful block validation. Nodes that are not present in in-memory trie will be read from disk.
This will help to validate state root without saving storage modifications to disk.