Add a CLI for Syncing (#548)
* Add a CLI for Syncing

This patch begins to add a CLI for syncing the blockchain. We create a new sub-app `CLI` to host command-line functions, which is currently limited to syncing. Over time, we will expand this CLI to sync, as well as host an RPC server, etc. We allow syncing from HTTP or IPC (e.g. Infura or a local node).

* Move option parser to separate tested file

* Fix issues with Credo

* Fix typespec for function by giving it real return

* Fix typespec for add_block_to_tree

* Alias all the things

* Fix broken alias

* Improvements to pass chain down from CLI arguments

* Continue to add more tests and documentation

* Fix issue with syncing uncle blocks

* Fix documentation for updated sync CLI

* Add significant number of test cases and improve overall design

* Fix minor Credo issue

* Use a simple progress bar for now

* Fix IPC syncing

* Refactor get_best_block to use loaded block, not to reload it

* Remove now deprecated sync with infura script

* Move more documentation to the readme