Extract `run_common_tests` for Block and Transaction Test (#547)
* Extract `run_common_tests` for Block and Transaction Test

The goal of this patch is to re-unify `run_common_tests` through all of our test cases. Previously, we generated a `test do...` for each test, which caused a number of issues. `BlockchainTest` opted instead of run a single Elixir test that creates async tasks for each test and aggregates the results at the end. This patch seeks to move the rest of the testcases on to this style, which is significantly faster overall. We also start to extract "known_failures" to a .txt file so we have a clear indication of failing tests in a limited set of places.

* Extract functions

* Fix credo

* Fix dialyzer

* Fix order of arguments