Adds Ethash dataset and cache sizes
This is the first step in adding the Ethash algorithm for Proof-of-work.
We follow the implementation outlined in Appendix J of the Yellow Paper
(Byzantium edition).

We introduce `dataset_size`, `cache_size` and `epoch` functions outlined
in J.2. We also introduce cached sizes for the first 2048 epochs. This
data comes from https://github.com/ethereum/wiki/wiki/Ethash#data-sizes
and significantly speed up the retrieval of those sizes.

Co-authored-by: German Velasco <germsvel@gmail.com>