Implement Ethash.seed_hash
This implements the `seed_hash/1` function for the `Ethash` algorithm.
This particular function can be found in Appendix J, section J.3.1 in
the Byzantium Edition of the Yellow Paper.

It's worth noting that previous editions of the yellow paper had the
original seed hash as the `keccak(<<0::256>>)` but Geth and Parity has
simply used `<<0::256>>` (not the keccak) so that became the canon and
is reflected in the Byzantium version of the yellow paper.

See the related issue: https://github.com/ethereum/yellowpaper/issues/312