Add Ethash.initial_cache (#555)
Adds the initial cache calculation to `Ethash` algorithm. This is found
in Appendix J, section J.3.2, of the Yellow Paper.

The initial cache is an array of of size (cache_size/j_hashytes) where
each element is calculated based on the previous element. Each element
is encrypted with keccak 512, so we expose that function from
ExthCrypto.