Add build-essential for debian/ubuntu installation
I am not a debian/ubuntu user, but when running the `bin/setup` script
in a remote box, the installation failed once because `make` was not
available and another time because `c++` was not available (when
compiling `rox`).

To fix this, I know we can add the `build-essential` package which
includes `make`, `gcc`, and `g++`. I don't believe the same package is
available for fedora/centos, so I did not include that package for them.