Switch group_to_module to pattern match
Why:

Currently, group_to_module is concatenating two strings and turning the
result into an atom to get from an atom to a module. Profiling reveals
that we do this a lot in the course of running code on the EVM.

This PR:

Switches to use pattern matching from known atoms to their corresponding
modules. As a benchmarking example, this takes the fibonacci16
performance test from the common tests from 34_000 milliseconds to
28_000 milliseconds.