Fix DER Encoding Bugs
This patch addresses a code quality issue that has continually bit us: a DER-encoded public key is `<<0x04, public_key:binary()>>`. The problem with that is, we often rely on the variable name to know whether or not the key was DER encoded or is the raw public key material. This has led to us using DER-encoded keys as standard keys or getting argument errors not giving DER-encoded keys to the `:crypto` module (whereas most Ethereum functions require non-DER encoded keys). We opt to now store DER-encoded keys as `{:public_der, public_key}` so that we cannot possibly pass the value incorrectly to a function (as dialyzer will catch those issues). This direct relates to us having a bug on this topic.