Start to be smarter about bootnodes and syncing
This patch starts getting our syncing code smarter about what peers it's connected to and why. We allow a user to specify bootnodes from the command line, and we check to see if we're connected to any nodes before we request new blocks. If we haven't yet connected, we'll wait and retry. This logic will grow as we start to run nodes that are consistently connected to 10-25 peers.