Add Fake Kademlia GenServer for Tests (#597)
A few tests relied on having a global Kademlia.Server gen server running for the tests to pass. This made these tests very difficult as any changes to configuration caused them to fail, as well as issues with running tests asynchronously. This patch creates a simple fake Kademlia Server that makes sure that each test starts this server as it needs it to reduce global dependencies.