Minor Re-Factor to UDP Module (#596)
This patch simply changes the UDP gen server to accept multiple arguments to `init` instead of a tuple. This is to help us with type-checking since the params hash is difficult to understand when or when not the options will be provided. This change is made in relation to several code points passing incorrect values that were not caught by dialyzer.