Fix Network Tests (#599)
This patch fixes our network tests. The goal is to make sure the tests run well against local peers or the bootnodes from the chains. We also make sure the network tests can be run locally or in CircleCI. Though, we opt not to run the tests in CI (this can be changed by making a branch name that ends in "network"). The fixes include:

1. Allowing an external IP address to be specified to give a remote node a real peer id with an IP address that is not 127.0.0.1.
2. Fix a discovery test that initiates find neighbours and checks that we've sent pings. We opt to ignore the pong responses as a simple way of tracking which nodes we've contracted so far.
3. We don't run remote connection tests asynchronously, since it meant we sent multiple connection attempts simultaneously to the same peer.
4. We increase timeouts, since real nodes take some time to respond.