Support IPv6 Addresses (#607)
* Support IPv6 Addresses

This patch allows us to support peers with IPv6 addresses. Erlang has good tools on parsing addresses, though not as much as determining if this might be an address, so when we parse a peer node, we first try it as an IP address, and if that fails, we return it as a charlist (domain name). This allows us to connect to peers with IPv6 addresses, which, for instance, is the enode address when you run `geth --nat none`.

* Fix for dialyzer