Fix Peer Sync (#613)
* Fix Peer Sync

This patch fixes peer sync (syncing against a single node). Previously, once we pulled 100 blocks, we would just stop and do nothing. Now, we correctly continue syncing whenever we have no blocks queued up. While we should be smarter about which blocks to sync and when to ask for more and from which peers, we'll probably end up switching to fast or warp sync before long, which might render this moot. With this patch, syncing against a local parity node works very well (and so does geth before it drops us for not having fast sync enabled).

* Clean up functions

* Fix minor test issues

* Fix dialyzer issue