Add proof-of-work algorithm
Why:

We would like to be able to verify the results given by other nodes when
adding a new block that the required work had been done to do so.

This PR:

Implements the Proof-of-work algorithm described in Appendix J of the
yellow paper (Byzantium edition).

Note: We've only implemented the full version and have added a pending
test for the light version. This will be implemented in a follow up PR.

Co-authored-by: German Velasco <germsvel@gmail.com>