Optimize Ethash cache generation in two places (#665)
Why?
===

When attempting to run the ethereum common tests for the Ethash PoW, it
became apparent that the cache generation is slow.

Note: the tests are not included in this commit because they are not yet
passing since they take too long.

What?
=====

This commit includes two optimizations:

1. The initial cache generation (based on a seed hash) recurses N
times, and each element is the Keccak 512 of the previous element.
Previous to this commit, we were starting with the first element and
going to element N, without keeping an accumulated list of elements. So
element N would generate all previous elements only for its own
calculation. Now we simply keep the values so that we only run through
the initial cache generation step N times. (where N is
cache_size/@j_hasbytes).

2. When performing the RandMemoHash algorithm, we were using an elixir
List since that is the format of the cache. But we must constantly
update that list as part of the RandMemoHash algorithm. Since getting
and updating elements of a linked list is slow, we now transform the
cache into a Map with the indexes as keys. This allows for faster
accessing and updating of the cache.

Co-authored-by: German Velasco <germsvel@gmail.com>