Allow Failure from Disconnected Peer
If a peer asks to disconnect, it probably will terminate the connection immediately after asking to disconnect. By the time we get to processing that message, we try to disconnect to an already disconnected peer, which causes the connection manager to crash (since we're matching :ok when trying to disconnect to a disconnected peer). As such, we simply allow any result from disconnect, since it will disconnect or it's been disconnected.