Stream Batch Updates for Trie
Previously, when we had a large number of updates, we would send them all to the NIF at once. For a several megabyte update (which is common in warp sync), this would effectively crash the erlang runtime system (since NIF cannot be pre-empted). This patch changes the system to instead stream the updates in batches. The allows us to play nice with pre-emption, allows for faster and smaller garbage collection and keeps the overall memory footprint significantly lower.