Add Exth Helper App
Currently, we have a variety of functions (like Ethereum specific helpers for encoding/decoding hex, etc.) that do not have a home, or live in the project where they happened to be first used. This creates a lot of repeat code that ends up being potentially bugging. Here, we create a new app, Exth, which can be used for a variety of shared utilities. Currently, we simply have some debug functions and hex encoders, but we could move in functions from other areas that do not necessarily belong anywhere else. Again, these functions are specific to Ethereum/mana, but not to any other sub-project in particular.

Bump cache

Remove old