Optimize Ethash proof-of-work (#670)
Using `Enum.at/2` is expensive with lists. We have been using lists
throughout the Ethash module because the Yellow Paper and other
implementations tend to use arrays. But since we have to access
and update those lists constantly, we update the Cache to be a map with
indexes for keys. This allows for faster reads and inserts.

When profiling, we found that one of the hotspots was `FNV.hash_lists/2`.
We update that function to no longer use `Enum.at` but to recursively
walk through each pair of elements instead.

Co-authored-by: Matthew Sumner <matt.m.sumner@gmail.com>