Complete Ethash implementation
Description
===========

This commit fixes a few issues that we encountered with the Ethash
implementation. The issues were raised when trying to pass the PoW
Ethereum common tests.

Some of the issues fixed:

- Use 32 bit unsigned integer in RandMemoHash algorithm (generates
correct cache)

- Use correct 32 bit unsigned integer in mix initialization

Testing
-------

Note that we only test the light version of the ethash algorithm. The
full version simply generates the full dataset prior to calculating the
algorithm. The light version lazily generates the dataset items it
requires.

We only test the light version because of time constraints. The dataset
generation for the PoW common tests takes a long time, so we do not want
to introduce that into CI.

When validating block headers or mining, we would likely store the cache
and the dataset in memory since it only changes every epoch.

Co-authored-by: Matthew Sumner <matt.m.sumner@gmail.com>