Add Support for all Parity Chains (#671)
* Add Support for all Parity Chains

This patch simply adds basic support for all chains that are supported for parity. As we aren't currently verifying proof of work nor authority, it's fairly simple to add new chains. In the future, to properly enable these chains, we will need to make sure we can handle the given engine (POW or POA). Also, even if we do not support a chain, this means we could write a nice message like "We do not support Ellacoin due to ...".

* Use pattern matching

* Add newline at the EOFs