Add Block `with_hash` and `fetch_block_hash` helpers
The Block struct contains information about a block, and the block hash is a very important and commonly used field. However, as it's based on the other data in a block (which may have been added peicemeal), it's not obvious when we should calculate the blockhash (though it's very obvious when we need it). This makes block_hash a good candidate for lazy evaluation (calculate the value when it's needed, but keep it around for later uses). We standardize this by adding two functions `Block.with_hash` and `Block.fetch_block_hash` which handle reading the block hash if it exists and calculating it if it doesn't. As Elixir values are immutable, we don't have a good way of forcing downstream code to continue to pass around the struct with the calculated value, but we can try our best that it does.